pyproject.toml
[project]
name = "porcupy"
version = "0.1.2"
description = "Crested Porcupine Optimizer (CPO) with GPU acceleration for optimization problems"
authors = [
    {name = "Samman Sarkar", email = "sarkar.samman4231@gmail.com"}
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.7"
dependencies = [
    "numpy>=1.19.0",  # Increased minimum version for better GPU compatibility
    "scipy>=1.4.0"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Mathematics",
]

[project.optional-dependencies]
plotting = ["matplotlib>=3.3.0"]
# GPU acceleration support (CUDA required)
cuda = ["cupy-cuda11x"]  # Default to CUDA 11.x
cuda11x = ["cupy-cuda11x"]
cuda12x = ["cupy-cuda12x"]
# Development dependencies with GPU support
gpu-dev = [
    "cupy-cuda11x",
    "pytest>=6.0.0",
    "pytest-cov>=2.10.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=3.8.0",
]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.10.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=3.8.0",
]
all = [
    "matplotlib>=3.3.0",
    "pytest>=6.0.0",
    "pytest-cov>=2.10.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=3.8.0",
    # Include CUDA support in 'all' but make it optional
    # Users should install with: pip install -e ".[all,cuda]" for GPU support
]


[project.urls]
Homepage = "https://github.com/SammanSarkar/porcupy"
Documentation = "https://github.com/SammanSarkar/porcupy/tree/main/docs"
Issues = "https://github.com/SammanSarkar/porcupy/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=porcupy --cov-report=term-missing"